requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.0
langchain-openai==0.0.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
slack-sdk==3.26.2
python-dotenv==1.0.0
pandas==2.1.4
tabulate==0.9.0
